+++ src/main/java/ladder/domain/result/Reward.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.result;
 
 public class Reward {
