+++ src/test/java/ladder/domain/ladder/LadderPropertyTest.java
@@ -1,7 +1,8 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import ladder.domain.ladder.LadderProperty;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
