+++ src/test/java/ladder/domain/ladder/LadderTest.java
@@ -1,5 +1,7 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
+import ladder.domain.ladder.Ladder;
+import ladder.domain.ladder.Line;
 import ladder.utils.BooleanGenerator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
