+++ src/test/java/ladder/domain/player/PlayerTest.java
@@ -1,7 +1,10 @@
-package ladder.domain;
+package ladder.domain.player;
 
 import static org.assertj.core.api.Assertions.*;
 
+import ladder.domain.player.Name;
+import ladder.domain.player.Player;
+import ladder.domain.player.StartIndex;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
