+++ src/test/java/ladder/domain/player/StartIndexTest.java
@@ -1,7 +1,8 @@
-package ladder.domain;
+package ladder.domain.player;
 
 import static org.assertj.core.api.Assertions.*;
 
+import ladder.domain.player.StartIndex;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
