+++ src/test/java/ladder/domain/result/ResultTest.java
@@ -1,8 +1,10 @@
-package ladder.domain;
+package ladder.domain.result;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import ladder.domain.result.Result;
+import ladder.domain.result.Reward;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
