+++ src/test/java/ladder/domain/result/ResultsTest.java
@@ -1,10 +1,12 @@
-package ladder.domain;
+package ladder.domain.result;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 
+import ladder.domain.result.Result;
+import ladder.domain.result.Results;
+import ladder.domain.result.Reward;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
